professor teaching

+++ ETA/model.py
@@ -150,7 +150,7 @@ class Model(tf_keras.Model):
 
     def q_update_val(self, loss):
 
-        epsilon = 0.9 / (10 * self.gcounter)
+        epsilon = 0.9 / tf.cast(10 * self.gcounter, tf.float32)
 
         action = tf.cond(
             tf.logical_or(
